arrumei o erro do banco

+++ database/migrations/2025_10_06_114425_add_website_to_users_table.php
@@ -4,16 +4,18 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
+class AddWebsiteToUsersTable extends Migration
 {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('website')->nullable();
-        });
+        if (!Schema::hasColumn('users', 'website')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->string('website')->nullable();
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('website');
-        });
+        if (Schema::hasColumn('users', 'website')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('website');
+            });
+        }
     }
 };
